+++ NYTimesMostViewedArticles/DetailScreen/ArticleDetailViewModel.swift
@@ -7,16 +7,6 @@
 
 import Foundation
 
-protocol UIViewControllerType: AnyObject {
-    func push(viewController: UIViewControllerType, animated: Bool)
-}
-
-protocol ArticleRepresentable {
-    var title: String? { get }
-    var author: String? { get }
-    var date: String? { get }
-}
-
 protocol ArticleDetailView: UIViewControllerType {
     var viewModel: ArticleDetailViewModel! { get set }
     func configureUI()
